# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/webhooks": {
      "get": {
        "description": "Every subscription on your account, deleted ones excluded. Each one says which events it receives and where they are delivered.",
        "operationId": "WebhookController_getWebhooks_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Your subscriptions.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookListResponseDTO"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or revoked API key."
          }
        },
        "security": [
          {
            "x-api-key": []
          }
        ],
        "summary": "List your webhook subscriptions",
        "tags": [
          "webhooks"
        ]
      },
      "post": {
        "description": "textbee POSTs the events you pick to your delivery URL and signs each request with your signing secret, sent as the X-Signature header. Failed deliveries are retried up to 10 times, and a subscription that keeps failing is paused.",
        "operationId": "WebhookController_createWebhook_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWebhookDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The subscription was created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "The delivery URL is unusable, the signing secret is shorter than 20 characters, the event list is empty or unknown, or you already have the maximum number of subscriptions."
          },
          "401": {
            "description": "Missing, invalid, or revoked API key."
          }
        },
        "security": [
          {
            "x-api-key": []
          }
        ],
        "summary": "Create a webhook subscription",
        "tags": [
          "webhooks"
        ]
      }
    },
    "/api/v1/webhooks/notifications": {
      "get": {
        "description": "Delivery history across your subscriptions, newest first. Use it to see what textbee sent, what your endpoint answered, and whether a retry is pending. History is kept for subscriptions you have since deleted.",
        "operationId": "WebhookController_getWebhookNotifications_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page to return. Default 1.",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Records per page. Default 10.",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Only return deliveries in this state.",
            "schema": {
              "enum": [
                "pending",
                "retrying",
                "delivered",
                "failed"
              ],
              "type": "string"
            }
          },
          {
            "name": "eventType",
            "required": false,
            "in": "query",
            "description": "Only return deliveries for this event.",
            "schema": {
              "enum": [
                "MESSAGE_RECEIVED",
                "MESSAGE_SENT",
                "MESSAGE_DELIVERED",
                "MESSAGE_FAILED",
                "UNKNOWN_STATE",
                "SMS_STATUS_UPDATED"
              ],
              "type": "string"
            }
          },
          {
            "name": "deviceId",
            "required": false,
            "in": "query",
            "description": "Only return deliveries for messages from this device.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start",
            "required": false,
            "in": "query",
            "description": "Start of the time range, ISO 8601. Applied only together with end.",
            "schema": {
              "example": "2026-08-01T00:00:00Z",
              "type": "string"
            }
          },
          {
            "name": "end",
            "required": false,
            "in": "query",
            "description": "End of the time range, ISO 8601.",
            "schema": {
              "example": "2026-08-31T23:59:59Z",
              "type": "string"
            }
          },
          {
            "name": "webhookSubscriptionId",
            "required": false,
            "in": "query",
            "description": "Only return deliveries for one subscription.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of delivery records.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookNotificationListResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "deviceId or webhookSubscriptionId is not a valid id."
          },
          "401": {
            "description": "Missing, invalid, or revoked API key."
          },
          "404": {
            "description": "No webhook subscription with that id on your account."
          }
        },
        "security": [
          {
            "x-api-key": []
          }
        ],
        "summary": "List webhook delivery attempts",
        "tags": [
          "webhooks"
        ]
      }
    },
    "/api/v1/webhooks/{webhookId}": {
      "get": {
        "description": "One subscription with its delivery counters.",
        "operationId": "WebhookController_getWebhook_v1",
        "parameters": [
          {
            "name": "webhookId",
            "required": true,
            "in": "path",
            "description": "Subscription id, from GET /webhooks.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The subscription.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookResponseDTO"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or revoked API key."
          },
          "404": {
            "description": "No webhook subscription with that id on your account."
          }
        },
        "security": [
          {
            "x-api-key": []
          }
        ],
        "summary": "Get a webhook subscription",
        "tags": [
          "webhooks"
        ]
      },
      "patch": {
        "description": "Changes only the fields you send. Use isActive to re-enable a subscription textbee paused after repeated failures.",
        "operationId": "WebhookController_updateWebhook_v1",
        "parameters": [
          {
            "name": "webhookId",
            "required": true,
            "in": "path",
            "description": "Subscription id, from GET /webhooks.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWebhookDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated subscription.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "The delivery URL is unusable, the signing secret is shorter than 20 characters, or the event list is empty or unknown."
          },
          "401": {
            "description": "Missing, invalid, or revoked API key."
          },
          "404": {
            "description": "No webhook subscription with that id on your account."
          }
        },
        "security": [
          {
            "x-api-key": []
          }
        ],
        "summary": "Update a webhook subscription",
        "tags": [
          "webhooks"
        ]
      },
      "delete": {
        "description": "Stops all deliveries for this subscription. Its delivery history stays readable through GET /webhooks/notifications.",
        "operationId": "WebhookController_deleteWebhook_v1",
        "parameters": [
          {
            "name": "webhookId",
            "required": true,
            "in": "path",
            "description": "Subscription id, from GET /webhooks.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The subscription was deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookDeletedResponseDTO"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or revoked API key."
          },
          "404": {
            "description": "No webhook subscription with that id on your account."
          }
        },
        "security": [
          {
            "x-api-key": []
          }
        ],
        "summary": "Delete a webhook subscription",
        "tags": [
          "webhooks"
        ]
      }
    }
  },
  "info": {
    "title": "textbee API",
    "description": "Send and receive SMS through an Android phone you own. Authenticate every request with an API key from your textbee dashboard, sent as the x-api-key header.",
    "version": "1.0",
    "contact": {},
    "license": {
      "name": "MIT",
      "url": "https://github.com/textbee/textbee/blob/main/LICENSE"
    }
  },
...
  "servers": [
    {
      "url": "https://cloudtextapi.frionode.online"
    }
  ],
  "components": {
    "securitySchemes": {
      "x-api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key from your textbee dashboard, sent on every request."
      }
    },
    "schemas": {
...
      "PaginationMetaDTO": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "description": "Current page number"
          },
          "limit": {
            "type": "integer",
            "description": "Number of items per page"
          },
          "total": {
            "type": "integer",
            "description": "Total number of items"
          },
          "totalPages": {
            "type": "integer",
            "description": "Total number of pages"
          }
        },
        "required": [
          "page",
          "limit",
          "total",
          "totalPages"
        ]
      },
...
      "WebhookNoteDTO": {
        "type": "object",
        "properties": {
          "at": {
            "format": "date-time",
            "type": "string",
            "description": "When the note was added."
          },
          "text": {
            "type": "string",
            "description": "What textbee recorded, for example why the subscription was paused."
          }
        },
        "required": [
          "at",
          "text"
        ]
      },
      "WebhookSubscriptionDTO": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Subscription id."
          },
          "user": {
            "type": "string",
            "description": "Owner account id."
          },
          "name": {
            "type": "string",
            "description": "Label shown in the dashboard."
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether deliveries are attempted. textbee pauses subscriptions that keep failing."
          },
          "events": {
            "type": "array",
            "description": "Events this subscription receives.",
            "items": {
              "type": "string",
              "enum": [
                "MESSAGE_RECEIVED",
                "MESSAGE_SENT",
                "MESSAGE_DELIVERED",
                "MESSAGE_FAILED",
                "UNKNOWN_STATE",
                "SMS_STATUS_UPDATED"
              ]
            }
          },
          "deliveryUrl": {
            "type": "string",
            "description": "URL events are POSTed to."
          },
          "signingSecret": {
            "type": "string",
            "description": "Secret used to sign deliveries."
          },
          "successfulDeliveryCount": {
            "type": "number",
            "description": "Deliveries that succeeded."
          },
          "deliveryFailureCount": {
            "type": "number",
            "description": "Deliveries that failed."
          },
          "deliveryAttemptCount": {
            "type": "number",
            "description": "Delivery attempts made, retries included."
          },
          "lastDeliveryAttemptAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last time a delivery was attempted."
          },
          "lastDeliverySuccessAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last time a delivery succeeded."
          },
          "lastDeliveryFailureAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last time a delivery failed."
          },
          "notes": {
            "description": "Notes textbee added, such as an auto-pause reason.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WebhookNoteDTO"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "When the subscription was created."
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "When it was last updated."
          }
        },
        "required": [
          "_id",
          "user",
          "isActive",
          "events",
          "deliveryUrl",
          "signingSecret",
          "successfulDeliveryCount",
          "deliveryFailureCount",
          "deliveryAttemptCount",
          "createdAt",
          "updatedAt"
        ]
      },
      "WebhookListResponseDTO": {
        "type": "object",
        "properties": {
          "data": {
            "description": "Your webhook subscriptions. Deleted ones are not returned.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WebhookSubscriptionDTO"
            }
          }
        },
        "required": [
          "data"
        ]
      },
      "WebhookNotificationDTO": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Delivery record id."
          },
          "webhookSubscription": {
            "type": "string",
            "description": "Subscription this belongs to."
          },
          "event": {
            "type": "string",
            "enum": [
              "MESSAGE_RECEIVED",
              "MESSAGE_SENT",
              "MESSAGE_DELIVERED",
              "MESSAGE_FAILED",
              "UNKNOWN_STATE",
              "SMS_STATUS_UPDATED"
            ],
            "description": "Event that was delivered."
          },
          "payload": {
            "type": "object",
            "description": "Exact JSON body textbee POSTed to your endpoint."
          },
          "computedStatus": {
            "type": "string",
            "enum": [
              "pending",
              "retrying",
              "delivered",
              "failed"
            ],
            "description": "Delivery state. Retrying means textbee will try again, failed means it gave up after 10 attempts."
          },
          "deliveryUrl": {
            "type": "string",
            "description": "URL the delivery was sent to when it was attempted."
          },
          "deliveryAttemptCount": {
            "type": "number",
            "description": "Attempts made for this event."
          },
          "deliveredAt": {
            "format": "date-time",
            "type": "string",
            "description": "When your endpoint accepted the delivery."
          },
          "lastDeliveryAttemptAt": {
            "format": "date-time",
            "type": "string",
            "description": "When the last attempt was made."
          },
          "nextDeliveryAttemptAt": {
            "format": "date-time",
            "type": "string",
            "description": "When the next retry is due."
          },
          "deliveryAttemptAbortedAt": {
            "format": "date-time",
            "type": "string",
            "description": "When textbee stopped retrying."
          },
          "errorType": {
            "type": "string",
            "enum": [
              "retryable",
              "non-retryable"
            ],
            "description": "Whether the failure is worth retrying."
          },
          "httpStatusCode": {
            "type": "number",
            "description": "Status code your endpoint returned on the last attempt."
          },
          "responseBody": {
            "type": "string",
            "description": "First 1000 characters of your endpoint response."
          },
          "idempotencyKey": {
            "type": "string",
            "description": "Idempotency key sent with the delivery."
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "When the event was recorded."
          }
        },
        "required": [
          "_id",
          "webhookSubscription",
          "event",
          "payload",
          "computedStatus",
          "deliveryAttemptCount",
          "createdAt"
        ]
      },
      "WebhookNotificationPageDTO": {
        "type": "object",
        "properties": {
          "data": {
            "description": "Delivery records, newest first.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WebhookNotificationDTO"
            }
          },
          "meta": {
            "description": "Pagination info.",
            "allOf": [
              {
                "$ref": "#/components/schemas/PaginationMetaDTO"
              }
            ]
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "WebhookNotificationListResponseDTO": {
        "type": "object",
        "properties": {
          "data": {
            "description": "Delivery records with pagination.",
            "allOf": [
              {
                "$ref": "#/components/schemas/WebhookNotificationPageDTO"
              }
            ]
          }
        },
        "required": [
          "data"
        ]
      },
      "WebhookResponseDTO": {
        "type": "object",
        "properties": {
          "data": {
            "description": "The subscription.",
            "allOf": [
              {
                "$ref": "#/components/schemas/WebhookSubscriptionDTO"
              }
            ]
          }
        },
        "required": [
          "data"
        ]
      },
      "CreateWebhookDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Label shown in the dashboard. Up to 64 characters.",
            "example": "Order notifications"
          },
          "deliveryUrl": {
            "type": "string",
            "description": "URL textbee POSTs events to. Must be http or https and publicly reachable. Private and loopback hosts are rejected.",
            "example": "https://example.com/textbee/webhook"
          },
          "signingSecret": {
            "type": "string",
            "description": "Shared secret, at least 20 characters. textbee signs every delivery with it and sends the signature in the X-Signature header, so your endpoint can verify the request really came from textbee."
          },
          "events": {
            "type": "array",
            "description": "Events to deliver. At least one is required.",
            "example": [
              "MESSAGE_RECEIVED"
            ],
            "items": {
              "type": "string",
              "enum": [
                "MESSAGE_RECEIVED",
                "MESSAGE_SENT",
                "MESSAGE_DELIVERED",
                "MESSAGE_FAILED",
                "UNKNOWN_STATE",
                "SMS_STATUS_UPDATED"
              ]
            }
          }
        },
        "required": [
          "deliveryUrl",
          "signingSecret",
          "events"
        ]
      },
      "UpdateWebhookDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Label shown in the dashboard. Up to 64 characters.",
            "example": "Order notifications"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether deliveries are attempted. Set it to true to re-enable a subscription textbee paused after repeated failures."
          },
          "deliveryUrl": {
            "type": "string",
            "description": "New delivery URL. Same rules as on create: http or https, no private or loopback hosts.",
            "example": "https://example.com/textbee/webhook"
          },
          "signingSecret": {
            "type": "string",
            "description": "New signing secret, at least 20 characters."
          },
          "events": {
            "type": "array",
            "description": "Replacement event list. Cannot be empty.",
            "example": [
              "MESSAGE_RECEIVED",
              "MESSAGE_DELIVERED"
            ],
            "items": {
              "type": "string",
              "enum": [
                "MESSAGE_RECEIVED",
                "MESSAGE_SENT",
                "MESSAGE_DELIVERED",
                "MESSAGE_FAILED",
                "UNKNOWN_STATE",
                "SMS_STATUS_UPDATED"
              ]
            }
          }
        }
      },
      "WebhookDeletedResultDTO": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the delete succeeded."
          }
        },
        "required": [
          "success"
        ]
      },
      "WebhookDeletedResponseDTO": {
        "type": "object",
        "properties": {
          "data": {
            "description": "Outcome of the delete.",
            "allOf": [
              {
                "$ref": "#/components/schemas/WebhookDeletedResultDTO"
              }
            ]
          }
        },
        "required": [
          "data"
        ]
      }
    }
  }
}